added enpoint meteors, response in JSON format outputed to enpoint response
